pyproject.toml
[project]
name = "bearcart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy<2.0.0",
    "opencv-python>=4.11.0.86",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "rshell>=0.0.36",
    "torch>=2.7.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.22.0",
]
